feat: Add class to store and manipulate datasets
Add a class for storing and manipulating an array of objects of a
specific class.

The Dataset class relies on DataType having a unique attribute called
'id'. Some problems arise from this: for instance, a Person class may
have its own 'id' attribute which conflicts with how Dataset uses the
property.

There are ways around this, such as wrapping each object in a new
'DataRow' class and setting a unique index/id attribute there. Given the
scope and intent of the project however, I have decided it is not worth
addressing for the time being.

note: findIndex() returns -1 when it can't find an item, which splice()
treats as a valid, negative index. To prevent the last array item from
being removed, check if index is valid before calling splice().