refactor: Return UI components directly instead of via object property
Since there is no reason here to attach components to an object, it is
more straightforward to return and access UI components directly.

For cases where more than one component is returned (such as a modal and
its opening button), return an array which can be split into separate
variables.