feat: Store data using localStorage
Use localStorage to store todo/project state across browser sessions -
This required some changes to how the Project and Todo classes were
implemented.

The first change is how id is assigned: both class constructors now use
an 'id' parameter, so ids can be retained instead of being randomly
generated each time they are loaded in (this is especially important for
Project ids, since that's how Todos link themselves to a Project).

The second change was adding a toJSON method to ensure all relevant
properties are included when objects are JSON.stringify(ed), which
ensures getter/setter properties are not ignored.