fix: Stop TodoRow from throwing errors when loading todos without a date
Originally, it was assumed todo.date was a valid date string, but it can
also be empty (i.e., todo with no date). This would cause format to try
and parse non-existant dates (resulting in an error).

To fix this, only format the date if it exists - otherwise, display an
empty string.