fix: Fix format for datetime input in EditTodoModal
Previously, 'hh' was passed into the format function to generate hours
for dateString. However, using lowercase 'hh' formats hours in a
[1-12] hour format, instead of the [0-23] hour format expected by
datetime-local.

This caused the datetime-local input to display the incorrect value for
time values, specifically values that should have been PM. To fix this,
use uppercase 'HH' within the format function to generate hours in the
expected [0-23] hour format.