fix: Fix processing of search data in SearchBar
This fix resolves two related issues: 1) Incorrect usage of the forEach
method, and 2) inconsistent use of toLowerCase.

In javascript, primitives behave as if "passed by value" in functions.
Meaning, when forEach passes a value from dataArr to the input
function, modifying that value has no effect on the original value.

This in turn makes attempts to change dataArr using forEach do nothing,
including trying to convert each string to lowercase. To fix this, use
the map method to generate a new array with modified values and return
that instead.

What made this bug hard to find was I somehow missed two variables
(projectName and projectDescription) in dataArr that called toLowerCase
in their initial assignment, making it appear that forEach "worked" for
some values but not others.

This led to a long chain of trial and error messing with values that
"weren't working," until I noticed the actual error. You live and you
learn, I suppose.