feat: Rework search bar to update immediately on user input
The previous "change" event fired when the search bar input was changed,
but not necessarily for every change. Relevant here is it does not fire
when the user types something in to the input, meaning enter must be
pressed to actually trigger a search.

The "input" event fires when the input has been changed "as a direct
result of a user action" (see docs for "input" below). Relevant here is
this includes typing.

Use "input" to trigger updates to the queryFn instead of "change". This
means results are displayed/updated as soon as the user starts typing,
making things more immediate and intuitive.

See:
https://developer.mozilla.org/en-US/docs/Web/API/Element/input_event
https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/change_event